move form intent and add flow to view model
