+++ form/src/main/java/org/dhis2/form/ui/intent/FormIntent.kt
@@ -1,4 +1,4 @@
-package org.dhis2.form.ui
+package org.dhis2.form.ui.intent
 
 import org.dhis2.form.mvi.MviIntent
 
@@ -9,11 +9,8 @@ sealed class FormIntent : MviIntent {
         val month: Int,
         val day: Int
     ) : FormIntent()
-    data class SelectDateFromYearMonthDayAgeCalendar(
-        val uid: String,
-        val date: String?
-    ) : FormIntent()
 
     data class OpenCustomAgeCalendar(val uid: String, val label: String) : FormIntent()
-    data class SelectDateFromCustomAgeCalendar(val uid: String, val date: String?) : FormIntent()
+    data class SelectDateFromAgeCalendar(val uid: String, val date: String?) : FormIntent()
+    data class ClearDateFromAgeCalendar(val uid: String) : FormIntent()
 }
